Move common parts in handlers, refactor
